app/requirements.txt
Flask==1.1.2
Flask-JWT-Extended==3.25.0
Flask-Injector
python-dotenv==0.15.0
injector==0.18.4
boto3
Pillow
